Uncomment and update CMake build settings
The `include_directories` and `link_directories` commands in CMakeList.txt were uncommented to make them functional. Discussions and personal notes about the build process for Apple and Windows, as well as the success of the build, have been marked as a multiline comment, ensuring they don't interfere with code execution.