Update documentation and clean up main.cpp
This commit improves the clarity of the project documentation by moving guide for adding functionalities to docs/README.md from README.md. Meanwhile, minor refactoring is done in main.cpp, focusing on renaming variables and tidying up the event loop for consistency and readability. The README.md file is also updated to clarify the current status of the project.