Add EventSystem with Application, Mouse, and Key events
This commit introduces the EventSystem to the application and includes the addition of ApplicationEvent, MouseEvent, and KeyEvent classes. In addition, a new md file named 'EventDispatcher.md' is added to the docs/Architecture/EventSystem for planning the event dispatcher system. On the other hand, an outdated file 'SampleMarkdownDemo.md' is deleted from the repository.