Update code documentation and integrate Doxygen
Updated and added code documentation while integrating Doxygen for auto-generation of documentation. The modifications primarily provide additional explanations and details for classes and functions. Several enhancements were also made which include the removal of the getGLFWwindow() function in GenericWindow.cpp, and setting the RECURSIVE tag in Doxyfile to YES, allowing examination of files in subdirectories.