- Included glad library in ImGuiOpenGLRenderer.h and ImGuiLayer.cpp - Modified gitignore file to ignore Debug folder - Updated OpenGL initialization in ImGuiLayer and GenericWindow
In this commit, glad library is included in both ImGuiOpenGLRenderer.h and ImGuiLayer.cpp to improve GUI support. Debug folder is now ignored in the project to avoid unnecessary clutter in version control. Additionally, some changes are made to the OpenGL initialization in ImGuiLayer and GenericWindow to enhance the app's graphical compatibility and performance.