- Commented out LayerTower::get().onUpdate() in Application.cpp - Added glfwSwapBuffers in GenericWindow::onUpdate()
In an effort to optimize the application's performance, the update function for the LayerTower class was commented out within Application.cpp. Furthermore, to ensure that the latest display data gets shown, the glfwSwapBuffers function call was added to the GenericWindow class's onUpdate method. These adjustments are expected to improve the overall functionality of the application.