- Implemented InputLayer and InputState classes - Added InputLayer integration into Application - Adjusted LayerStack to be managed by Application instead of singleton - Updated ImGuiLayer to implement getAcceptingEventFlags method
This commit involves significant updates to revamp application layering systems. This includes the introduction of InputLayer and InputState classes, which handle input events and state across the application, and the integration of the new InputLayer into the main Application class. This commit also includes adjustments to LayerStack management, moving from a singleton model to being managed directly by the Application. Furthermore, the ImGuiLayer class has been updated to include an implementation for the getAcceptingEventFlags method to better handle event flags.