- Added ImGuiLayer, TemplateLayer, and DebugLayer classes to Ayla::Core::Layers - Renamed namespace Ayla to Ayla::Core in Application.h - Moved key and mouse event handler to InputState - Expanded KeyCodes enum in InputState.h - Changed Window namespace to Ayla::Windows - Likewise, the Event namespace is now Ayla::Events
In this update, the main focus was on redefining various namespaces for better organization and specificity, the implementation of new layer classes, and the expansion of key codes and event handling within the InputState class. Improvements to the InputLayer class and the Ayla::Core::Layers namespace are aimed at enhancing debugging and program flexibility.