** Fixed window size getter ** : WindowProperties was not getting updated by the window resize callback.
- Changed time output from milliseconds to seconds in Application.cpp
- Added display size output in ImGuiLayer.cpp
- Modified WindowProperties default size in Window.h
- Refactored GenericWindow to add nodiscard attribute in GenericWindow.h
- Updated output and cursors in update method in GenericWindow.cpp

The commit includes several formatting and documentation improvements. It changes the time display units in Application.cpp and adds display size output in ImGuiLayer.cpp. Additionally, the default size of WindowProperties and the nodiscard attribute in the GenericWindow class are updated. Lastly, improved outputs and cursor positions are incorporated in the update method of GenericWindow.cpp file.