- Deleted ImGuiEvents.cpp and ImGuiEvents.h files - Improved KeyCode support in KeyState class and renamed Keycode files - Disabled automatic Google Test in CMakeLists.txt - Added and improved some functionalities in InputLayer and InputState classes - Converted IsKeyDown function calls to isKeyDown in ImGui Classes
The IsKeyDown function calls were converted to isKeyDown in ImGui classes for better consistency, and the Google Test was disabled in the CMakeLists.txt file to prevent it from running automatically. Some functions in InputLayer and InputState classes were also added and enhanced. In the KeyCode class, support for translating and handling keys was improved. Furthermore, the ImGuiEvents.cpp and ImGuiEvents.h files were removed due to redundancy. This message was generated by AI.