** SAVE POINT 2 (EVERYTHING ACTUALLY WORKS !!) ** - Introduced condition to display ID Stack Tool Window - Changed variable name for clarity
In the ImGuiLayer.cpp file, the code for displaying the ID Stack Tool Window is now only run if the 'showStackTool' variable is set to true. This adds an additional level of control. Furthermore, the variable for showing the demo window has been renamed for better readability and understanding of its function.