- Refactored 'Event' classes to 'IEvent' - Updated functions across multiple files to use 'IEvent' instead of 'Event' - Revised several event methods to include 'nodiscard' attribute - Modified the naming from 'update' to 'Update' and 'renderImGui' to 'RenderImGui' - Added or enhanced several comments for methods and classes
This commit involves a major refactor of the 'Event' classes to 'IEvent' which affects a number of files. This change is made to align with Interface naming conventions. As part of this update, multiple class functions have been modified to utilize 'IEvent' instead of 'Event'. The 'nodiscard' attribute has been added to several event methods to ensure the return value is being used as intended. Function names have been modified to follow consistent naming conventions. Comments for various classes and methods have also been added or improved for clarity.